rename and add print
