+++ python/msa/plugins/conversation/module.py
@@ -1,6 +1,6 @@
 
-from msa.plugins.conversation_module import event
-from msa.plugins.conversation_module import coroutine
+from msa.plugins.conversation import event
+from msa.plugins.conversation import coroutine
 from msa.module import Module
 
 class PluginModule(Module):
